backend/requirements.txt
# VibeWeather Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# HTTP
httpx==0.28.0

# AI / LLM
openai==1.57.0

# ML
xgboost==2.1.3
numpy==2.1.3
scikit-learn==1.5.2   # preprocessing için

# Validation
pydantic==2.10.3

# Utils
python-dotenv==1.0.1